+++ src/modules/block/components/BlockResultList/BlockResultList.css
