+++ src/modules/block/components/BlockResultList/BlockResultList.js
