+++ src/modules/block/components/BlockResultList/index.js
