+++ src/modules/transaction/components/TransactionResultList/TransactionResultList.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import Icon from '@basics/icon';
 import { truncateAddress } from '@wallet/utilities/account';
+import { getTxConfig } from '@search/utils';
 import LiskAmount from '@shared/liskAmount';
 import styles from './TransactionResultList.css';
-import { getTxConfig } from '../../utils';
 
 const Transactions = ({
   t, transactions, onSelectedRow, rowItemIndex, updateRowItemIndex, activeToken,
